requirements.txt
langchain
chainlit
unstructured[local-inference]
chromadb
pinecone-client
tiktoken
openai
gradio
layoutparser[layoutmodels,tesseract]